feat: add assert.rs; remove subtraction
