+++ manta-crypto/src/eclair/uint.rs
@@ -15,11 +15,12 @@
 // along with manta-rs.  If not, see <http://www.gnu.org/licenses/>.
 
 //! Unsigned Integers Implementation
+
 use crate::{
-    constraint::{Add, Sub},
+    constraint::Add,
     eclair::{
-        bool::AssertWithinRange,
-        ops::{AddAssign, SubAssign},
+        assert::AssertWithinRange,
+        ops::AddAssign,
     },
 };
 
@@ -38,12 +39,12 @@ impl<T> U128<T> {
     }
 }
 
-impl<T, COM> Add<Self, COM> for U128<T>
+impl<O, T, COM> Add<Self, COM> for U128<T>
 where
-    COM: AssertWithinRange<T, 128>,
-    T: Add<T, COM, Output = T>,
+    COM: AssertWithinRange<O, 128>,
+    T: Add<T, COM, Output = O>,
 {
-    type Output = Self;
+    type Output = O;
 
     #[inline]
     fn add(self, rhs: Self, compiler: &mut COM) -> Self::Output {
@@ -59,34 +60,6 @@ where
     #[inline]
     fn add_assign(&mut self, rhs: Self, compiler: &mut COM) {
         self.0.add_assign(rhs.0, compiler);
-        compiler.assert_within_range(&self.0);
-    }
-}
-
-impl<T, COM> Sub<Self, COM> for U128<T>
-where
-    COM: AssertWithinRange<T, 128>,
-    T: Sub<T, COM, Output = T>,
-{
-    type Output = Self;
-
-    // TODO: Should we check self.0 > rhs.0 before subtraction?
-    // TODO: Should we check self > rhs instead of `AssertWithinRange`?
-    #[inline]
-    fn sub(self, rhs: Self, compiler: &mut COM) -> Self::Output {
-        Self::new(self.0.sub(rhs.0, compiler), compiler)
-    }
-}
-
-impl<T, COM> SubAssign<Self, COM> for U128<T>
-where
-    COM: AssertWithinRange<T, 128>,
-    T: SubAssign<T, COM>,
-{
-    // TODO: Should we check self.0 > rhs.0 before subtraction?
-    #[inline]
-    fn sub_assign(&mut self, rhs: Self, compiler: &mut COM) {
-        self.0.sub_assign(rhs.0, compiler);
         compiler.assert_within_range(&self.0);
     }
 }
